fix some bugs;add color;fix super display of registered cloud;super read obj to do

+++ include/ICPViewer.h
@@ -75,7 +75,7 @@ public Q_SLOTS:
   randomButtonPressed ();
 
   void
-  distance_ValueChanged (double value);
+  distance_ValueChanged (int value);
 
   void
   npoints_ValueChanged (int value);
